+++ src/main/java/de/intevation/lada/rest/OrtService.java
@@ -14,9 +14,12 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.UriInfo;
 
+import de.intevation.lada.authentication.Authentication;
+import de.intevation.lada.authentication.AuthenticationException;
 import de.intevation.lada.data.QueryBuilder;
 import de.intevation.lada.data.Repository;
 import de.intevation.lada.model.Ort;
@@ -32,6 +35,10 @@ public class OrtService
     @Named("ortrepository")
     private Repository repository;
 
+    @Inject
+    @Named("ldapauth")
+    private Authentication authentication;
+
     /**
      * Request a LZusatzWert via its id.
      *
@@ -41,8 +48,19 @@ public class OrtService
     @GET
     @Path("/{id}")
     @Produces("text/json")
-    public Response findById(@PathParam("id") String id) {
-        return repository.findById(Ort.class, id);
+    public Response findById(
+        @PathParam("id") String id,
+        @Context HttpHeaders headers
+    ) {
+        try {
+            if (authentication.isAuthorizedUser(headers)) {
+                return repository.findById(Ort.class, id);
+            }
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
+        catch(AuthenticationException ae) {
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
     }
 
     /**
@@ -55,34 +73,69 @@ public class OrtService
      */
     @GET
     @Produces("text/json")
-    public Response filter() {
-        return repository.findAll(Ort.class);
+    public Response filter(@Context HttpHeaders headers) {
+        try {
+            if (authentication.isAuthorizedUser(headers)) {
+                return repository.findAll(Ort.class);
+            }
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
+        catch(AuthenticationException ae) {
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
     }
 
     @PUT
     @Produces("text/json")
     @Path("/{ortId}")
     @Consumes("application/json")
-    public Response update(Ort ort) {
-        return repository.update(ort);
+    public Response update(Ort ort, @Context HttpHeaders headers) {
+        try {
+            if (authentication.isAuthorizedUser(headers)) {
+                return repository.update(ort);
+            }
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
+        catch(AuthenticationException ae) {
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
     }
 
     @POST
     @Produces("text/json")
     @Consumes("application/json")
-    public Response create(Ort ort) {
-        return repository.create(ort);
+    public Response create(Ort ort, @Context HttpHeaders headers) {
+        try {
+            if (authentication.isAuthorizedUser(headers)) {
+                return repository.create(ort);
+            }
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
+        catch(AuthenticationException ae) {
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
     }
 
     @DELETE
     @Path("/{ortId}")
-    public Response delete(@PathParam("ortId") String ortId ) {
-        Response response = repository.findById(Ort.class, ortId);
-        Ort ort = (Ort)response.getData();
-        if (ort != null) {
-            repository.delete(ort);
-            return new Response(true, 200, null);
+    public Response delete(
+        @PathParam("ortId") String ortId,
+        @Context HttpHeaders headers
+    ) {
+        try {
+            if (authentication.isAuthorizedUser(headers)) {
+                Response response = repository.findById(Ort.class, ortId);
+                Ort ort = (Ort)response.getData();
+                if (ort != null) {
+                    repository.delete(ort);
+                    return new Response(true, 200, null);
+                }
+                return new Response(false, 600, null);
+            }
+            return new Response(false, 699, new ArrayList<Ort>());
         }
-        return new Response(false, 600, null);
+        catch(AuthenticationException ae) {
+            return new Response(false, 699, new ArrayList<Ort>());
+        }
     }
 }
